corpus-utils creates corpus taming mappings
The corpora contain a number of unhelpful rare and stray characters, many
of which can be mapped together to reduce the alphabet size.

This will create a python file containing the mapping in a dictionary.

    ./corpus-utils -m corpus/whatever -d --decompose-caps > \
         mappings/whatever.py

I wasn't initially going to do it this way, rather create the mapping at
run time, but printing it out is useful to see what's going on.